Comunicação Front-Back | Adição de Subtarefa

+++ back-end/front-end/js/tarefas.js
@@ -1,7 +1,9 @@
 let dadosProjeto;
 let dadosSessao;
 const form = document.getElementById("task-form");
-let editando = 0; // para saber se estamos editando ou criando
+const formSub = document.getElementById("sub-form");
+let editando = 0; // para saber se estamos editando ou criando a tarefa
+let editandoSub = 0; // para saber se estamos editando ou criando a subtarefa
 let idTarefaEditar;
 
 document.getElementById("add-task").addEventListener("click", () => {
@@ -14,9 +16,11 @@ function formatarDataISOparaBR(dataISO) {
 }
 
 function abrirFormulario() {
-  document.getElementById("form-container").style.display = "flex";
-  document.getElementById("form-title").innerHTML = "Nova Tarefa";
-  document.getElementById("label-anexo").innerHTML = "Anexo da Tarefa";
+  if ((dadosProjeto.projeto.id_gestor === dadosSessao.dados.id || dadosProjeto.projeto.ids_administradores.includes(dadosSessao.dados.id)) && (dadosProjeto.projeto.status !== "Concluído")) {
+    document.getElementById("form-container").style.display = "flex";
+    document.getElementById("form-title").innerHTML = "Nova Tarefa";
+    document.getElementById("label-anexo").innerHTML = "Anexo da Tarefa";
+  }
 }
 
 function fecharFormulario() {
@@ -37,7 +41,7 @@ async function editarTarefa(idTarefa) {
 
   const dados = await resposta.json();
 
-  if (dados.result){
+  if (dados.result) {
     document.getElementById("form-container").style.display = "flex";
     document.getElementById("form-title").innerHTML = "Editar Tarefa";
     document.getElementById("label-anexo").innerHTML = "Substituir Anexo da Tarefa";
@@ -47,10 +51,10 @@ async function editarTarefa(idTarefa) {
     document.getElementById("data").value = dataFormatada;
     document.getElementById("descricao").value = dados.tarefa.descricao;
     editando = 1;
-  }else{
+  } else {
     alert(dados.mensagem);
   }
-  
+
 }
 
 async function alterarTarefa() {
@@ -105,46 +109,54 @@ function corStatus(status) {
   if (status === "atrasada") return "vermelho";
 }
 
-async function moverParaCima(idTarefa, ordem) {
-  if (ordem === 1){
-    return;
-  }
+async function moverParaCima(idTarefa) {
+
+  const tarefaAtual = document.getElementById(idTarefa);
+  const tarefaAnterior = tarefaAtual.previousElementSibling;
+
+  if (!tarefaAnterior) return;
 
   const resposta = await fetch('http://localhost:8080/tarefas/updateOrdem/' + idTarefa, {
-        method: 'PUT',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ tipo: 'antecipar' })
-    });
+    method: 'PUT',
+    credentials: 'include',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ tipo: 'antecipar' })
+  });
 
   const dados = await resposta.json();
 
-  if (!dados.result){
-    alert(dados.mensagem);
-  }else{
-    window.location.reload();
+  if (!dados.result) {
+    return alert(dados.mensagem);
   }
+
+  // Troca no DOM
+  tarefaAtual.parentNode.insertBefore(tarefaAtual, tarefaAnterior);
+  return;
 }
 
-async function moverParaBaixo(idTarefa, ordem, qtTarefas) {
-  if (ordem === qtTarefas){
-    return;
-  }
+async function moverParaBaixo(idTarefa) {
+
+  const tarefaAtual = document.getElementById(idTarefa);
+  const tarefaPosterior = tarefaAtual.nextElementSibling;
+
+  if (!tarefaPosterior) return;
 
   const resposta = await fetch('http://localhost:8080/tarefas/updateOrdem/' + idTarefa, {
-        method: 'PUT',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ tipo: 'regredir' })
-    });
+    method: 'PUT',
+    credentials: 'include',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ tipo: 'regredir' })
+  });
 
   const dados = await resposta.json();
 
-  if (!dados.result){
-    alert(dados.mensagem);
-  }else{
-    window.location.reload();
+  if (!dados.result) {
+    return alert(dados.mensagem);
   }
+
+  // Troca no DOM
+  tarefaPosterior.parentNode.insertBefore(tarefaPosterior, tarefaAtual);
+  return;
 }
 
 async function encerrarSessao() {
@@ -183,15 +195,15 @@ async function baixarAnexoTarefa(nomeArquivo) {
 async function carregarDados() {
 
   const dadosUsuario = await fetch('http://localhost:8080/usuarios/verificaSessao/true', {
-        method: 'GET',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'include'
-    });
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+    credentials: 'include'
+  });
 
   dadosSessao = await dadosUsuario.json();
 
   if (!dadosSessao.result) {
-      return window.location.href = "login.html";
+    return window.location.href = "login.html";
   }
 
   const urlAtual = window.location.href;
@@ -229,11 +241,12 @@ async function carregarDados() {
       for (i = 0; i < dadosTarefas.tarefas.length; i++) {
         const tarefa = dadosTarefas.tarefas[i];
         const card = document.createElement("section");
+        card.id = tarefa.id;
 
         let data;
-        if (tarefa.status === "Concluída"){
+        if (tarefa.status === "Concluída") {
           data = tarefa.data_entrega;
-        }else{
+        } else {
           data = tarefa.data_limite;
         }
 
@@ -243,13 +256,13 @@ async function carregarDados() {
 
         let descDataTarefa;
         let statusTarefa;
-        if (tarefa.status === "Concluída"){
+        if (tarefa.status === "Concluída") {
           statusTarefa = "verde";
           descDataTarefa = "Entregue em ";
-        }else if (tarefa.status === "Atrasada"){
+        } else if (tarefa.status === "Atrasada") {
           statusTarefa = "vermelho";
           descDataTarefa = "Data Máxima Entrega: ";
-        }else if (tarefa.status === "Pendente"){
+        } else if (tarefa.status === "Pendente") {
           statusTarefa = "azul";
           descDataTarefa = "Data Máxima Entrega: ";
         }
@@ -260,12 +273,19 @@ async function carregarDados() {
         }
 
         let exibirConcluirEditar = ``;
-        if (dadosProjeto.projeto.id_gestor === dadosSessao.dados.id || dadosProjeto.projeto.ids_administradores.includes(dadosSessao.dados.id)){
-          if (tarefa.status === "Concluída"){
+        let exibirMudarOrdem = ``;
+        let btnAddSubTarefa = ``;
+        if ((dadosProjeto.projeto.id_gestor === dadosSessao.dados.id || dadosProjeto.projeto.ids_administradores.includes(dadosSessao.dados.id)) && (dadosProjeto.projeto.status !== "Concluído")) {
+          if (tarefa.status === "Concluída") {
             exibirConcluirEditar = `<a class="btn-concluir" onclick="statusTarefa('${tarefa.id}', 'Reabrir')">Reabrir</a>`;
-          }else{
+            exibirMudarOrdem = `<button onclick="moverParaCima('${tarefa.id}')" title="Subir" class="position">🔼</button><button onclick="moverParaBaixo('${tarefa.id}')" title="Descer" class="position">🔽</button>`;
+          } else {
             exibirConcluirEditar = `<a class="btn-concluir" onclick="statusTarefa('${tarefa.id}', 'Concluir')">Concluir</a>&nbsp;<a onclick="editarTarefa('${tarefa.id}')" class="edit-tarefa">Editar</a>&nbsp;<a class="btn-excluir" onclick="excluirTarefa('${tarefa.id}')">Excluir</a>`;
+            exibirMudarOrdem = `<button onclick="moverParaCima('${tarefa.id}')" title="Subir" class="position">🔼</button><button onclick="moverParaBaixo('${tarefa.id}')" title="Descer" class="position">🔽</button>`;
+            btnAddSubTarefa = `<button onclick="abrirFormularioSubTarefa('${tarefa.id}')" class="subTarefa">ADICIONAR SUBTAREFA</button>`;
           }
+        }else{
+          document.getElementById('add-task').style.display = "none";
         }
 
         card.className = "categoria";
@@ -274,7 +294,7 @@ async function carregarDados() {
           <p><b>${descDataTarefa}</b> ${dataFormatada}</p><br>
           <span class="status ${statusTarefa}">${tarefa.status}</span><br><br>
           <p class="descricao">${tarefa.descricao}</p><br>
-          <button onclick="addSubTarefa('${tarefa.id}')" class="subTarefa">ADICIONAR SUBTAREFA</button>
+          ${btnAddSubTarefa}
           <div class="tarefas">
         `);
 
@@ -291,13 +311,13 @@ async function carregarDados() {
             let status;
             let descData;
 
-            if (subtarefa.status === "Concluída"){
+            if (subtarefa.status === "Concluída") {
               status = "verde";
               descData = "Entregue em ";
-            }else if (subtarefa.status === "Atrasada"){
+            } else if (subtarefa.status === "Atrasada") {
               status = "vermelho";
               descData = "Data Máxima Entrega: "
-            }else if (subtarefa.status === "Pendente"){
+            } else if (subtarefa.status === "Pendente") {
               status = "azul";
               descDataTarefa = "Data Máxima Entrega: ";
             }
@@ -307,9 +327,9 @@ async function carregarDados() {
             const dataFormatada = `${dia}/${mes}/${ano}`;
 
             let baixarAnexoSub;
-            if (subtarefa.anexo === null){
+            if (subtarefa.anexo === null) {
               baixarAnexoSub = "";
-            }else{
+            } else {
               baixarAnexoSub = `<a onclick="baixarAnexoSubTarefa('${subtarefa.anexo}')">Baixar Anexo</a>`;
             }
 
@@ -322,20 +342,20 @@ async function carregarDados() {
               <div style="margin-top: 10px;"><button onclick="editarSubTarefa('${subtarefa.id}')">Editar</button>${baixarAnexoSub}
               </div>
               <br>
-              <button onclick="moverParaCimaSub(${subtarefa.id}, ${subtarefa.ordem})" title="Subir">🔼</button><button onclick="moverParaBaixoSub(${subtarefa.id}, ${subtarefa.ordem}, ${dadosSubTarefas.subtarefas.length})" title="Descer">🔽</button>
+              <button onclick="moverParaCimaSub(${subtarefa.id})" title="Subir">🔼</button><button onclick="moverParaBaixoSub(${subtarefa.id})" title="Descer">🔽</button>
               </div>
             `);
 
           }
         }
-        
+
 
         card.insertAdjacentHTML('beforeend', `
           </div>
           <br><br>
-          <div class="btns-tarefa"><div>${exibirConcluirEditar}</div><br><div>${exibirAnexo}&nbsp;<button onclick="moverParaCima('${tarefa.id}', ${tarefa.ordem})" title="Subir" class="position">🔼</button><button onclick="moverParaBaixo('${tarefa.id}', ${tarefa.ordem}, ${dadosTarefas.tarefas.length})" title="Descer" class="position">🔽</button></div>
+          <div class="btns-tarefa"><div>${exibirConcluirEditar}</div><br><div>${exibirAnexo}&nbsp;${exibirMudarOrdem}</div>
         `);
-        
+
 
         containerTarefas.appendChild(card);
       }
@@ -352,7 +372,7 @@ async function carregarDados() {
 
 async function cadastrarTarefa() {
 
-  if (editando === 1){
+  if (editando === 1) {
     alterarTarefa();
     return;
   }
@@ -366,12 +386,14 @@ async function cadastrarTarefa() {
 
   if (titulo === "") {
     msgAviso.innerHTML = "Preencha o campo Título!";
-    return;
-  } else if (descricao === "") {
-    msgAviso.innerHTML = "Preencha o campo Descrição!";
+    document.getElementById("titulo").focus();
     return;
   } else if (!data || data === null || undefined) {
     msgAviso.innerHTML = "Preencha o campo Data!";
+    return;
+  } else if (descricao === "") {
+    msgAviso.innerHTML = "Preencha o campo Descrição!";
+    document.getElementById("descricao").focus();
     return;
   }
 
@@ -401,10 +423,10 @@ async function cadastrarTarefa() {
 
 async function statusTarefa(idTarefa, tipoAlteracao) {
 
-  if (!confirm(`Deseja realmente ${tipoAlteracao} essa tarefa?`)){
+  if (!confirm(`Deseja realmente ${tipoAlteracao} essa tarefa?`)) {
     return;
   }
-  
+
   const resposta = await fetch('http://localhost:8080/tarefas/updateStatus/' + idTarefa, {
     method: 'PUT',
     credentials: 'include',
@@ -415,12 +437,12 @@ async function statusTarefa(idTarefa, tipoAlteracao) {
   const dados = await resposta.json();
 
   if (dados.result) {
-    if (tipoAlteracao === "Concluir"){
+    if (tipoAlteracao === "Concluir") {
       alert("Tarefa concluida com sucesso!");
-    }else if (tipoAlteracao === "Reabrir"){
+    } else if (tipoAlteracao === "Reabrir") {
       alert("Tarefa reaberta com sucesso!");
     }
-  
+
     window.location.href = "tarefas.html?id=" + dadosProjeto.projeto.id;
   } else {
     alert(dados.mensagem);
@@ -429,7 +451,7 @@ async function statusTarefa(idTarefa, tipoAlteracao) {
 }
 
 async function excluirTarefa(idTarefa) {
-  if (!confirm('Deseja realmente excluir essa tarefa?')){
+  if (!confirm('Deseja realmente excluir essa tarefa?')) {
     return;
   }
 
@@ -441,12 +463,79 @@ async function excluirTarefa(idTarefa) {
 
   const dados = await resposta.json();
 
-  if (dados.result){
+  if (dados.result) {
     alert('Tarefa excluida com sucesso!');
     window.location.reload();
     return;
-  }else{
+  } else {
     alert('Erro: ' + dados.mensagem);
     return;
   }
+}
+
+function abrirFormularioSubTarefa(idTarefa){
+  if ((dadosProjeto.projeto.id_gestor === dadosSessao.dados.id || dadosProjeto.projeto.ids_administradores.includes(dadosSessao.dados.id)) && (dadosProjeto.projeto.status !== "Concluído")) {
+    document.getElementById("form-subtarefa").style.display = "flex";
+    document.getElementById("sub-title").innerHTML = "Nova Subtarefa";
+    document.getElementById("label-anexo").innerHTML = "Anexo da Subtarefa";
+    document.getElementById("id-tarefa").value = idTarefa;
+  }
+}
+
+function fecharFormularioSub(){
+  document.getElementById("form-subtarefa").style.display = "none";
+  formSub.reset();
+  editandoSub = 0;
+}
+
+async function cadastrarSubtarefa(){
+  const id_tarefa = document.getElementById("id-tarefa").value;
+  
+  if (editandoSub === 1) {
+    alterarSubTarefa();
+    return;
+  }
+
+  const titulo = document.getElementById('tituloSub').value.trim();
+  const data = document.getElementById('dataSub').value;
+  const descricao = document.getElementById('descricaoSub').value.trim();
+  const arquivo = document.getElementById('anexoSub').files[0];
+  const msgAviso = document.getElementById('msgAvisoSub');
+  msgAviso.innerHTML = "";
+
+  if (titulo === "") {
+    msgAviso.innerHTML = "Preencha o campo Título!";
+    document.getElementById("tituloSub").focus();
+    return;
+  } else if (!data || data === null || undefined) {
+    msgAviso.innerHTML = "Preencha o campo Data!";
+    return;
+  } else if (descricao === "") {
+    msgAviso.innerHTML = "Preencha o campo Descrição!";
+    document.getElementById("descricaoSub").focus();
+    return;
+  }
+
+  const formData = new FormData();
+  formData.append('anexo', arquivo);
+  formData.append('titulo', titulo);
+  formData.append('descricao', descricao);
+  formData.append('data_limite', data);
+  formData.append('id_tarefa', id_tarefa);
+
+  const resposta = await fetch('http://localhost:8080/subtarefas/', {
+    method: 'POST',
+    credentials: 'include',
+    body: formData
+  });
+
+  const dados = await resposta.json();
+
+  if (dados.result) {
+    alert("Subtarefa criada com sucesso!");
+    window.location.reload();
+  } else {
+    msgAviso.innerHTML = dados.mensagem;
+  }
+
 }
